Tercera entrega completada | Correccion de rutas absolutas

+++ js/utilidades.js
@@ -1,9 +1,16 @@
+
 import { salones } from "../constantes/salones.js";
 
 export function obtenerUsuarioLocalStorage() {
-  return JSON.parse(localStorage.getItem("usuario"));
+  return JSON.parse(localStorage.getItem("usuario")) || null;
 }
 
-(function inicializarLocalStorage() {
-  localStorage.setItem("salonesInfantiles", JSON.stringify(salones));
-})();
+function inicializarLocalStorage() {
+  const yaExiste = localStorage.getItem('salonesInfantiles')
+  if (!yaExiste) {
+    localStorage.setItem('salonesInfantiles', JSON.stringify(salones))
+
+  }
+}
+
+inicializarLocalStorage();
